Feature/dynamic catalog (#1)
Adds a functionality to load catalogs from an API call.

This PR will add a DynamicCatalogStore class which will be responsible for sending a GET request to an internal API and retrieve available catalog data on demand. From time to time, Presto will also send a GET request to get the delta of updated catalogs since last request.

* add data connection test payload

* add dynamic catalog store

* add data connection response

* adds data connection response object and tests the funcionality of retriving data from external api

* implements dynamic catalog store initial api configuration

* add api configuration file

* add dynamic catalog store files

* add explicit binder for DynamicCatalogStore

* call DynamicCatalogStore on Presto server load

* change sample config file to port 8081 - conflicts with api on 8080

* add catalog delta retriever scheduler

* add data connections delta update

* add unit tests to dynamic catalog store

* fix require non null checks